read zipcode boundary data from census dat files
Each ASCII file from the Census website contains boundaries for all the ZIP codes in a state. There are actually two files for each state: a metadata file that ends in 'a.dat' and the actual boundary data which ends in just '.dat'.

The boundary for each ZIP code has a single main polygon and may also contain several 'exclusions' such as islands, lakes, etc. This Python function will read in the data for a given state.